[kidash] Unbundle kidash completely
This commit unbundles kidash completely from grimoire-elk.
Now, all of kidash lives in the kidash directory, and
everything should be ready to relocate in a standalone repo.
The kidash.py binary is now kidash, in the bin directory.
The pip package is now kidash.
Version now is 0.4.0 (it was 0.30.x with GrimoireELK, but
then it was grimoire-kidash, so this is really a new package).