[import] Don't modify the height of visualizatons in dashboards from Kibana >= 6
If a dashboard has been created with Kibana >= 6 the issue with the
height=1 in which the visualization can be shown completly because
the title bar has been shown. So don't modify the height in these
cases becasue the dashboard creator has decided to use this heigth=1
in any case.

Fixes #5