Update mapping in Kibana6 if needed
When using Kibana 6, .kibana index mapping is set to strict. That means we cannot
include new fields in the index without explicitely add the corresponding mapping
for that field. This change detects when this error is raised, updates .kibana
mapping including `release_date`  field and tries to upload panel again.