Fix import/export of searches
`_id` checking for compatibility with old Kibana versions now
looks for type as prefix followed by `:` character. Old check
was wrong as it just checked if `_id` contained type in any
place, which is a problem because types are common words like
`search` that can be used as part of titles and ids as well.
Notice that old ids were not hashes, but string similar to
titles.