Add viz title to data source removal checks
GitLab visualizations weren't being added to Overview panel
when data source filtering is active. That was due to the use
of widget title in the checks, but not visualization title,
that is a different one. Widget's title is the one you set
in the dashboard, while the right one to use is the title you
set when creating the visualization, as it is hidden in the
dashboard and it doesn't change from dashboard to dashboard
as it is associated to the viz and not to the panel.

The old check using the title coming from the dashboard is
still there to avoid cases that could be already working based
on it.

Signed-off-by: Alberto Pérez García-Plaza <alpgarcia@bitergia.com>