[kidash] Replace release date comparison with object version
This code replaces the way two versions of the same
dashboard/index pattern are compared. Thus, the custom
field `release_date` is removed in favor of the standard
field `version`.
This change is needed since in Kibana versions greater
than 6.8 custom fields are not allowed in the .kibana
index. Thus this change allows to have compatible .kibanas
across different Kibana versions.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>