[kidash] Support import to Kibiter/Kibana 6.8
This code allows to import Sigils dashboards to
Kibana/Kibiter 6.8.x instances, granting backward
compatibilty with previous versions.

Kibana 6.8.x sets the .kibana mapping to strict,
thus it makes impossible to modify the .kibana index
after creation. For this reason, Kidash has been modified
in the following way.

- Kidash now uploads an ad-hoc .kibana mapping, which
contains the metadashboard and projectname types. This
is needed to guarantee that Kibiter top menu works for
Kibana 6.8.x

- The `release_date` attribute stored in .kibana for
versions prior to 6.8.x has been moved to a different
index: `.grimoirelab-sigils`. This is needed since custom
fields cannot be saved in Kibana standard objects anymore
(dashboard, index patterns, etc.). Thus, each time there
is an attempt to upload a dashboard or index pattern to
Kibana, the release_date is retrieved from this new index
and compared with the release_date included in the
uploaded objects. If the object can be uploaed, the
release date included into it is popped out and saved
to the `.grimoirelab-sigils` index together with the
id and type of the object.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>